+++ packages/terra-form-select/src/_constants.js
@@ -1,0 +1,23 @@
+const KeyCodes = {
+  BACKSPACE: 8,
+  TAB: 9,
+  ENTER: 13,
+  ESCAPE: 27,
+  SPACE: 32,
+  END: 35,
+  HOME: 36,
+  LEFT_ARROW: 37,
+  UP_ARROW: 38,
+  RIGHT_ARROW: 39,
+  DOWN_ARROW: 40,
+};
+
+const Variants = {
+  COMBOBOX: 'combobox',
+  DEFAULT: 'default',
+  MULTIPLE: 'multiple',
+  TAG: 'tag',
+  SEARCH: 'search',
+};
+
+export { KeyCodes, Variants };