+++ packages/terra-site/src/examples/avatar/AvatarImage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import Avatar from 'terra-avatar';
 import exampleAvatarImage from '../../assets/150x150.jpg';
-import Avatar from '../../../../../packages/terra-avatar/src/Avatar';
 
 const AvatarImage = () => (
   <Avatar ariaLabel="John Doe" image={exampleAvatarImage} alt="placeholder" />