+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.4.0 - (May 23, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.3.0 - (May 16, 2018)
 ------------------
 ### Changed