+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.12.0 - (May 23, 2018)
+------------------
+### Changed
+* Dependency on terra-modal removed, now depends on terra-abstract-modal.
+
 2.11.0 - (May 16, 2018)
 ------------------
 ### Changed