+++ packages/terra-form-field/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Changed nightwatch to webdriverIO tests
 
+2.11.0 - (May 23, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.10.0 - (May 16, 2018)
 ------------------
 ### Changed