+++ packages/terra-icon/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Replaced nightwatch tests with wdio tests.
+
+2.10.0 - (May 23, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.9.0 - (May 16, 2018)
 ------------------