+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,14 @@ ChangeLog
 Unreleased
 ----------
 
+3.2.0 - (May 23, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* New variables for theming
+
 3.1.0 - (May 16, 2018)
 ------------------
 ### Changed