+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Replaced nightwatch tests with wdio
+
+2.11.0 - (May 23, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.10.0 - (May 16, 2018)
 ------------------